move function 'dump_memory' to hardware dir
